Decide for foo.rs approach
